Beyond Reality Face Trial
This is a trial of virtual mirror with the help of BRF SDK for adobe AIR
for Desktop. The link for the SDK is
https://www.beyond-reality-face.com/
Add the src and the src_3dengine_extensions to references and the media
folder from the SDK's bin to the project's bin folder if you are doing
it the first time. Adding references is necessary but the media folder
is already added.